+++ src/clients/models.py
@@ -37,7 +37,11 @@ class Client(SQLModel, table=True):
 
     date_modification: datetime = Field(
         default_factory=lambda: datetime.now(UTC),
-        sa_column=Column(DateTime(timezone=True), onupdate=lambda: datetime.now(UTC)),
+        sa_column=Column(
+            DateTime(timezone=True),
+            default=lambda: datetime.now(UTC),
+            onupdate=lambda: datetime.now(UTC),
+        ),
     )
     date_desactivation: datetime | None = Field(default=None)